Multiple required changes and updates
