+++ src/App.css
@@ -55,13 +55,15 @@ input {
 }
 
 :root {
-    --primary-color: #282828;
+    --primary-color: #282b33;
 
-    --white: #fafafa;
+    --white: #efefef;
     --black: #1c1c1c;
     --dark-gray: #232323;
     --superlight-gray: #e5e5e5;
     --light-gray: #c9c9c9;
+    --light-text: #4b4b4b;
+    --divider: #cccccc;
 
     --php-color: #5DD7C1;
     --tip: #3d3e4d;
@@ -69,7 +71,23 @@ input {
     --error: #E55353;
     --horray: #40CC45;
     --challenge: #39A8E7;
-    --tooltip-color: #606060;
 }
 
+::-webkit-scrollbar {
+    width: 10px;
+}
 
+/* Track */
+::-webkit-scrollbar-track {
+    background: #f1f1f1;
+}
+
+/* Handle */
+::-webkit-scrollbar-thumb {
+    background: #b0b0b0;
+}
+
+/* Handle on hover */
+::-webkit-scrollbar-thumb:hover {
+    background: #555;
+}
