+++ src/components/courses-page/course-search-bar/CourseSearchBar.js
