+++ src/components/home/Home.css
@@ -1,13 +1,17 @@
 /* hero section */
+.home-page {
+    height: 100vh;
+    display: flex;
+    flex-direction: column;
+}
+
 .home-hero {
     display: flex;
     justify-content: center;
     align-items: center;
-    width: 100%;
-    min-height: 760px;
-    background-color: #eef3ff;
+    background-color: #f3f3f3;
     text-align: center;
-    padding: 0 2rem;
+    height: 100%;
     flex-direction: column;
     row-gap: 2rem;
 }
@@ -107,32 +111,4 @@
     100% {
         transform: translateY(0);
     }
-}
-
-/* lesson section */
-.lesson-section {
-    padding: 4rem 2rem;
-    display: flex;
-    flex-direction: column;
-    max-width: 1200px;
-    margin: auto;
-}
-
-.lessons-header {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    text-align: center;
-    max-width: 40rem;
-    margin: 0 auto 4rem;
-}
-
-.lessons-header p {
-    font-size: 1.4rem;
-}
-
-.lesson-widgets-container {
-    display: flex;
-    flex-wrap: wrap;
-    margin: 2rem 0 4rem 0;
 }
