+++ src/components/profile-page/ProfilePage.css
@@ -1,3 +1,48 @@
-.profile-page {
+.profile-page-stats-widgets {
+    justify-content: space-between;
+    display: flex;
+    flex-wrap: wrap;
+    gap: 2rem;
+    margin-bottom: 2rem;
+    margin-top: 1rem;
+}
 
+.profile-description {
+    max-width: 35rem;
+}
+
+.profile-description p {
+    margin: 0;
+    color: var(--light-text);
+}
+
+.save-button {
+    width: max-content;
+}
+
+.profile-field {
+    max-width: 50rem;
+    margin: 2rem 0;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+}
+
+.profile-field h4 {
+    color: var(--light-text);
+}
+
+.profile-field input, .profile-field textarea {
+    padding: 0.8rem 1rem;
+    width: 30rem;
+    font-size: 1rem;
+    border-radius: 8px;
+    border: 1px solid #e3e3e3;
+    color: var(--light-text);
+}
+
+.profile-field textarea {
+    height: 6rem;
+    resize: none;
+    outline: none;
 }
