+++ src/components/profile-page/stat-widget/StatWidget.css
@@ -1,0 +1,25 @@
+.stat-widget {
+    background-color: #EFEFEF;
+    min-width: 12rem;
+    padding: 0 1rem;
+    border-radius: 8px;
+    border-left: 8px solid var(--primary-color);
+    flex: 1;
+    transition: 300ms ease-in;
+    cursor: pointer;
+}
+
+.stat-widget:hover {
+    transform: scale(1.02);
+}
+
+.stat-widget p {
+    color: #5d5d5d;
+    font-size: 1rem;
+    margin-bottom: 1.2rem;
+}
+
+.big-stat {
+    line-height: 0.4;
+    font-size: 3rem;
+}
