+++ src/components/resuable/Divider/Divider.js
@@ -1,0 +1,7 @@
+export default function Divider(props) {
+    const height = props.height;
+    const width = props.width;
+    const color = props.color;
+    const margin = props.margin;
+    return (<div className="divider" style={{"height": height, "width": width, "backgroundColor": color, "margin": margin + " 0"}}/>)
+}
