+++ src/components/sidebar/category/Category.js
@@ -3,14 +3,15 @@ import "./Category.css"
 export default function Category(props) {
     const title = props.title;
     const items = props.items;
+    const pathName = window.location.pathname.replaceAll("/", "");
     return (<div className="category">
         <span className="category-title">{title}</span>
         <div className={"category-elements"}>
-            {items.map((item) => (
-                <a href={item.replaceAll(" ", "-").toLowerCase()} key={item} className={"category-subtitle"}>
+            {items.map((item) => (<a className={(item.replaceAll(" ", "-").toLowerCase() === pathName ? "active" : "") + " category-subtitle"}
+                                     href={item.replaceAll(" ", "-").toLowerCase()}
+                                     key={item}>
                     {item}
-                </a>
-            ))}
+                </a>))}
         </div>
     </div>)
 }
