+++ src/components/topbar/TopBar.css
@@ -14,9 +14,9 @@ header {
 .logo {
     font-size: 1.2rem;
     letter-spacing: 0.1rem;
+    color: var(--white);
     padding: 1.4rem 0;
     display: flex;
-    color: white;
 }
 
 .nav-items {
